Fixed a few issues with how errors were handled in the Sample REST app, moved samples into their own package.
