+++ controller/build.gradle
@@ -20,8 +20,26 @@ dependencies {
     compile group: 'javax.json', name: 'javax.json-api', version: '1.1.4'
 }
 
-mainClassName = 'org.spice.JdbcMySQLVersion'
+mainClassName = 'org.spice.sample.JdbcMySQLVersion'
 
 war {
     webAppDirName = "src/webapp"
+}
+
+def tomcatHome = System.getenv('CATALINA_HOME') ?: '/var/lib/tomcat9'
+
+task install(type: Copy) {
+    from war.archivePath
+    into "${tomcatHome}/webapps"
+
+    doLast {
+        println "ran install from ${war.archivePath} to ${tomcatHome}/webapps"
+    }
+}
+
+
+task my_test_task {
+    doLast {
+        println "${tomcatHome} ${war.archivePath}"
+    }
 }
